moved augment function in process_data.py to separate file cast outputs in mgcvae.py to long
