+++ experiments/nuScenes/augment_function.py
@@ -1,0 +1,8 @@
+import numpy as np
+
+
+def augment(scene):
+    scene_aug = np.random.choice(scene.augmented)
+    scene_aug.temporal_scene_graph = scene.temporal_scene_graph
+    scene_aug.map = scene.map
+    return scene_aug
